split big script into different files
